apps/usecases/requirements.txt
# Common packages for all our azurefed apps
python-dotenv
streamlit
semantic-kernel
msal
azure-storage-blob

# App sepcific packages
phidata
duckduckgo-search
newspaper4k
lxml_html_clean